requirements-pipeline.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.23.0
scipy>=1.10.0

# Data Collection
pandas-datareader>=0.10.0
yfinance>=0.2.28
fredapi>=0.5.1
requests>=2.31.0

# Apache Airflow
apache-airflow>=2.7.0
apache-airflow-providers-http>=4.5.0

# Data Version Control
dvc>=3.0.0
dvc[s3]>=3.0.0

# Data Validation & Quality
great-expectations>=0.18.0
pandas-profiling>=3.6.0
scipy>=1.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Alerts (optional - choose based on your needs)
# slack-sdk>=3.21.0  # For Slack alerts
# sendgrid>=6.10.0   # For email alerts